// src/primer-api/model/globalName.ts
/**
 * Generated by orval v6.27.1 🍺
 * Do not edit manually.
 * Primer backend API
 * A backend service implementing a pedagogic functional programming language.
 * OpenAPI spec version: 0.7
 */

export interface GlobalName {
  baseName: string;
  /** @minItems 1 */
  qualifiedModule: string[];
}